pyproject.toml
[project]
name = "fastapi-example"
version = "0.1.0"
description = "Example Service based on FastAPI"
authors = [
    {name = "Leonid Shepetovsky", email = "leonid.shepetovsky@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.6,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "toml (>=0.10.2,<0.11.0)",
    "pydantic (>=2.10.5,<3.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
